HIVE-28104. (2.3) Move HTTP related methods from Utils to HttpUtils in shims
