+++ hcatalog/webhcat/svr/src/main/java/org/apache/hive/hcatalog/templeton/Main.java
@@ -34,7 +34,7 @@ import org.apache.commons.lang3.ArrayUtils;
 import org.apache.hadoop.hive.common.classification.InterfaceAudience;
 import org.apache.hadoop.hive.common.classification.InterfaceStability;
 import org.apache.hadoop.hdfs.web.AuthFilter;
-import org.apache.hadoop.hive.shims.Utils;
+import org.apache.hadoop.hive.shims.HttpUtils;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.hadoop.security.authentication.client.PseudoAuthenticator;
 import org.apache.hadoop.security.authentication.server.PseudoAuthenticationHandler;
@@ -235,12 +235,12 @@ public class Main {
   public FilterHolder makeXSRFFilter() {
     String customHeader = null; // The header to look for. We use "X-XSRF-HEADER" if this is null.
     String methodsToIgnore = null; // Methods to not filter. By default: "GET,OPTIONS,HEAD,TRACE" if null.
-    FilterHolder fHolder = new FilterHolder(Utils.getXSRFFilter());
+    FilterHolder fHolder = new FilterHolder(HttpUtils.getXSRFFilter());
     if (customHeader != null){
-      fHolder.setInitParameter(Utils.XSRF_CUSTOM_HEADER_PARAM, customHeader);
+      fHolder.setInitParameter(HttpUtils.XSRF_CUSTOM_HEADER_PARAM, customHeader);
     }
     if (methodsToIgnore != null){
-      fHolder.setInitParameter(Utils.XSRF_CUSTOM_METHODS_TO_IGNORE_PARAM, methodsToIgnore);
+      fHolder.setInitParameter(HttpUtils.XSRF_CUSTOM_METHODS_TO_IGNORE_PARAM, methodsToIgnore);
     }
     FilterHolder xsrfFilter = fHolder;
